♻️ refactor: move collapseChatMessages to chatrender module
